+++ src/AssemblyInfo.cs
@@ -8,5 +8,5 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyCopyright("Copyright (c) NightOwlZzz / Owl")]
 [assembly: ComVisible(false)]
 [assembly: Guid("EA1BED5E-CE7C-49D0-BD86-3784A82B3F9A")]
-[assembly: AssemblyVersion("0.3.0.0")]
-[assembly: AssemblyFileVersion("0.3.0.0")]
+[assembly: AssemblyVersion("0.3.1.0")]
+[assembly: AssemblyFileVersion("0.3.1.0")]